perf: extract MemoizedMiniSpace to reduce re-renders
Extracted the individual board space rendering in `MiniMap` into a new `MemoizedMiniSpace` component. Wrapped it in `React.memo` with a custom equality function that checks `playersHere` and `propState`. During the dice roll animation, the `players` prop updates on every tick, which previously caused all 40 spaces to re-render. Now, only the specific spaces that the moving player enters or leaves will re-render, significantly reducing unnecessary DOM updates and improving animation smoothness.
